requirements.txt
# Core Dependencies (Python 3.11 compatible)
flask==2.3.3
flask-session==0.5.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-bcrypt==1.0.1
pandas==2.1.4
numpy==1.26.2
openpyxl==3.1.2
xlrd==2.0.1
matplotlib==3.8.0
seaborn==0.12.2
plotly==5.18.0
jinja2==3.1.2
werkzeug==2.3.7
python-dotenv==1.0.0
typing-extensions==4.8.0
xlsxwriter==3.1.9
watchdog==3.0.0
PyMySQL==1.1.0
cryptography==41.0.7
pyinstaller==6.18.0